+++ scripts/generate_match_patterns.py
@@ -1,5 +1,11 @@
 #!/usr/bin/env python
 
+import sys
+
+if sys.version_info[:2] < (3, 14):
+    print('Cannot run this on Python version < 3.14!')
+    sys.exit(1)
+
 from fst.astutil import *
 from fst import *
 from fst.astutil import FIELDS, AST_BASES
@@ -10,13 +16,13 @@ m_patterns = []
 for ast_cls in FIELDS:
     if ast_cls is Constant:
         m_patterns.append('''
-class MConstant(M_Pattern):
+class MConstant(MAST):
     ast_cls = Constant
 
     def __init__(
         self,
-        value: object,
-        kind: object = ...,
+        value: _Pattern,  # Ellipsis here is not a wildcard but a concrete value to match
+        kind: _Pattern = ...,
     ) -> None:
         self._fields = fields = ['value']
         self.value = value
@@ -30,7 +36,7 @@ class MConstant(M_Pattern):
 
     name = ast_cls.__name__
     fields = ast_cls._fields
-    args = ''.join(f'\n        {f}: object = ...,' for f in fields)
+    args = ''.join(f'\n        {f}: _Pattern = ...,' for f in fields)
     set_ = ''.join(f'\n\n        if {f} is not ...:\n            self.{f} = {f}\n            fields.append({f!r})' for f in ast_cls._fields)
 
     if args:
@@ -39,7 +45,7 @@ class MConstant(M_Pattern):
         init = ''
 
     m_patterns.append(f'''
-class M{name}(M_Pattern):
+class M{name}(MAST):
     ast_cls = {name}{init}
 '''.strip())
 
@@ -47,7 +53,7 @@ for ast_cls in AST_BASES:
     name = ast_cls.__name__
 
     m_patterns.append(f'''
-class M{name}(M_Pattern):
+class M{name}(MAST):
     ast_cls = {name}
 '''.strip())